backend/requirements.txt
Flask
pyshacl
sparqlwrapper
PyJWT
openai
gunicorn
